fix(emit): repair generator yield lint guard
