+++ crates/tsz-emitter/src/declaration_emitter/helpers/type_inference_return_normalization.rs
@@ -10,17 +10,6 @@ use tsz_parser::parser::syntax_kind_ext;
 use tsz_parser::parser::{NodeIndex, NodeList};
 use tsz_scanner::SyntaxKind;
 
-/// A single `yield` operand contribution to a generator's inferred yield type.
-struct YieldOperand {
-    /// The operand's declaration-emit type (fresh object/array operands already
-    /// have their member positions widened).
-    type_id: tsz_solver::types::TypeId,
-    /// Whether the operand is a fresh primitive literal (a bare literal token,
-    /// not a const/type assertion). A *single* fresh literal widens to its base;
-    /// a union of distinct literals is preserved.
-    is_fresh_literal: bool,
-}
-
 impl<'a> DeclarationEmitter<'a> {
     pub(in crate::declaration_emitter) fn function_body_preferred_return_type_text(
         &self,
@@ -428,281 +417,6 @@ impl<'a> DeclarationEmitter<'a> {
         let text = Self::parenthesize_first_generic_function_type_argument_text(&text);
         self.rewrite_current_source_named_import_type_text(&text)
             .unwrap_or(text)
-    }
-
-    pub(in crate::declaration_emitter) fn generator_yield_return_type_text(
-        &self,
-        is_async: bool,
-        body_idx: NodeIndex,
-    ) -> Option<String> {
-        let interner = self.type_interner?;
-        let generator_name = if is_async {
-            "AsyncGenerator"
-        } else {
-            "Generator"
-        };
-
-        let mut operands: Vec<YieldOperand> = Vec::new();
-        if !self.collect_generator_yield_operands(body_idx, is_async, &mut operands, 0) {
-            return None;
-        }
-
-        // No `yield` operands at all (an empty generator, or one whose only
-        // value-bearing path never yields). tsc infers `never` as the yield type
-        // — there is no value the generator can produce. (A bare `yield;`
-        // contributes `undefined` and is collected as an operand, so it does not
-        // reach this branch.)
-        if operands.is_empty() {
-            return Some(format!("{generator_name}<never, void, unknown>"));
-        }
-
-        // tsc computes the inferred yield type as
-        // `getWidenedType(getUnionType(<yield operand types>))`. The union
-        // regularizes the operand literal types, so a union of two or more
-        // *distinct* literals is preserved verbatim (`yield "x"; yield "y"` ->
-        // `"x" | "y"`; `yield 1; yield 2` -> `1 | 2`). Only a single fresh literal
-        // widens to its base (`yield "x"` -> `string`); a const-asserted or
-        // otherwise non-fresh single literal keeps its literal type. Fresh
-        // object/array operands already widened their member positions per
-        // operand in `collect_generator_yield_operands`.
-        // `union_literal_reduce` matches tsc's default `UnionReduction.Literal`
-        // for `getUnionType`: it preserves source order and absorbs literals into
-        // their primitive (`"x" | string` -> `string`) without removing structural
-        // subtypes — the same reduction the explicit-annotation union path uses.
-        let any_fresh_literal = operands.iter().any(|operand| operand.is_fresh_literal);
-        let union_ty =
-            interner.union_literal_reduce(operands.iter().map(|operand| operand.type_id).collect());
-
-        let yield_ty =
-            if any_fresh_literal && tsz_solver::query::is_literal_type(interner, union_ty) {
-                // A single fresh literal value across every yield — widen to its base.
-                tsz_solver::operations::widening::widen_literal_type(interner, union_ty)
-            } else {
-                union_ty
-            };
-
-        let yield_text = self.print_type_id_for_inferred_declaration(yield_ty);
-        if yield_text.is_empty() || yield_text == "any" {
-            return None;
-        }
-        Some(format!("{generator_name}<{yield_text}, void, unknown>"))
-    }
-
-    /// Collect the type of each `yield` operand in `node_idx`'s body for
-    /// declaration-emit yield-type inference. Returns `false` when an operand's
-    /// type cannot be resolved (the caller then keeps the conservative
-    /// `Generator<any, …>` fallback). Nested functions/classes are not descended
-    /// into — their yields belong to their own generator scope.
-    fn collect_generator_yield_operands(
-        &self,
-        node_idx: NodeIndex,
-        is_async: bool,
-        operands: &mut Vec<YieldOperand>,
-        depth: usize,
-    ) -> bool {
-        if node_idx.is_none() || depth > 128 {
-            return true;
-        }
-        let Some(node) = self.arena.get(node_idx) else {
-            return true;
-        };
-
-        match node.kind {
-            k if k == syntax_kind_ext::YIELD_EXPRESSION => {
-                let Some(yield_expr) = self.arena.get_unary_expr_ex(node) else {
-                    return false;
-                };
-                if yield_expr.asterisk_token {
-                    // `yield* <iterable>` contributes the *element* type of the
-                    // delegated iterable to the enclosing generator's inferred
-                    // yield type (e.g. `yield* [1, 2]` yields `number`). Resolve
-                    // that element type through the solver's iterator protocol so
-                    // delegated yields participate in declaration inference instead
-                    // of forcing the conservative `any` fallback.
-                    let Some(element) =
-                        self.yield_star_delegated_yield_type_id(yield_expr.expression, is_async)
-                    else {
-                        return false;
-                    };
-                    operands.push(YieldOperand {
-                        type_id: element,
-                        is_fresh_literal: false,
-                    });
-                    return true;
-                }
-                if yield_expr.expression.is_none() {
-                    // A bare `yield;` contributes `undefined`.
-                    operands.push(YieldOperand {
-                        type_id: tsz_solver::types::TypeId::UNDEFINED,
-                        is_fresh_literal: false,
-                    });
-                    return true;
-                }
-                match self.yield_operand_type(yield_expr.expression) {
-                    Some(operand) => {
-                        operands.push(operand);
-                        true
-                    }
-                    None => false,
-                }
-            }
-            k if k == syntax_kind_ext::FUNCTION_DECLARATION
-                || k == syntax_kind_ext::FUNCTION_EXPRESSION
-                || k == syntax_kind_ext::ARROW_FUNCTION
-                || k == syntax_kind_ext::CLASS_DECLARATION
-                || k == syntax_kind_ext::CLASS_EXPRESSION
-                || k == syntax_kind_ext::METHOD_DECLARATION
-                || k == syntax_kind_ext::CONSTRUCTOR
-                || k == syntax_kind_ext::GET_ACCESSOR
-                || k == syntax_kind_ext::SET_ACCESSOR =>
-            {
-                true
-            }
-            _ => self
-                .arena
-                .get_children(node_idx)
-                .into_iter()
-                .all(|child_idx| {
-                    self.collect_generator_yield_operands(child_idx, is_async, operands, depth + 1)
-                }),
-        }
-    }
-
-    /// Resolve a single (non-`yield*`, non-empty) `yield` operand to its
-    /// declaration-emit type, returning `None` when the type is unavailable or
-    /// `any`/`error` (so the caller keeps the conservative fallback).
-    ///
-    /// A fresh object/array literal operand widens its member positions up front
-    /// (`yield { a: 1 }` -> `{ a: number }`), mirroring `getWidenedType`, which
-    /// widens fresh objects regardless of how many operands the union has. A bare
-    /// primitive literal is reported as fresh so that a *single* such value widens
-    /// later while a union of distinct ones is preserved.
-    fn yield_operand_type(&self, expr_idx: NodeIndex) -> Option<YieldOperand> {
-        let interner = self.type_interner?;
-        let asserted = self.explicit_asserted_type_text(expr_idx).is_some();
-        let inner_idx = self
-            .skip_parenthesized_expression(expr_idx)
-            .unwrap_or(expr_idx);
-
-        // A bare (non-asserted) primitive literal yield operand. The checker
-        // stores the *widened* type at the yield site (`yield "x"` -> `string`),
-        // so reconstruct the precise fresh literal type from the AST and report it
-        // as fresh: a single such operand widens to its base, while a union of
-        // distinct ones is preserved (matching tsc's `getWidenedType(getUnionType)`).
-        if !asserted && let Some(literal_ty) = self.yield_literal_operand_type_id(inner_idx) {
-            return Some(YieldOperand {
-                type_id: literal_ty,
-                is_fresh_literal: true,
-            });
-        }
-
-        let type_id = self.get_node_type_or_names(&[expr_idx])?;
-        if type_id == tsz_solver::types::TypeId::ANY || type_id == tsz_solver::types::TypeId::ERROR
-        {
-            return None;
-        }
-        let type_id = self.widen_unique_symbol_value_type_for_dts(type_id, 0);
-
-        // A fresh object/array literal operand widens its member positions up
-        // front (`yield { a: 1 }` -> `{ a: number }`); tsc's `getWidenedType`
-        // widens fresh objects regardless of how many operands the union has.
-        let is_fresh_compound = self.arena.get(inner_idx).is_some_and(|node| {
-            node.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
-                || node.kind == syntax_kind_ext::ARRAY_LITERAL_EXPRESSION
-        });
-        let type_id = if is_fresh_compound {
-            tsz_solver::operations::widening::widen_literal_type(interner, type_id)
-        } else {
-            type_id
-        };
-
-        Some(YieldOperand {
-            type_id,
-            is_fresh_literal: false,
-        })
-    }
-
-    /// Reconstruct the precise (non-widened) literal type for a fresh primitive
-    /// literal yield operand directly from its AST node, returning `None` for any
-    /// non-literal expression. Used because the checker stores a widened type at
-    /// the yield site, which would erase the literal the union must preserve.
-    fn yield_literal_operand_type_id(
-        &self,
-        node_idx: NodeIndex,
-    ) -> Option<tsz_solver::types::TypeId> {
-        let interner = self.type_interner?;
-        let node = self.arena.get(node_idx)?;
-        match node.kind {
-            k if k == SyntaxKind::StringLiteral as u16
-                || k == SyntaxKind::NoSubstitutionTemplateLiteral as u16 =>
-            {
-                let literal = self.arena.get_literal(node)?;
-                Some(interner.literal_string(&literal.text))
-            }
-            k if k == SyntaxKind::NumericLiteral as u16 => {
-                let literal = self.arena.get_literal(node)?;
-                let value = literal
-                    .value
-                    .or_else(|| literal.text.replace('_', "").parse::<f64>().ok())?;
-                Some(interner.literal_number(value))
-            }
-            k if k == SyntaxKind::BigIntLiteral as u16 => {
-                let literal = self.arena.get_literal(node)?;
-                Some(interner.literal_bigint(&literal.text.replace('_', "")))
-            }
-            k if k == SyntaxKind::TrueKeyword as u16 => Some(interner.literal_boolean(true)),
-            k if k == SyntaxKind::FalseKeyword as u16 => Some(interner.literal_boolean(false)),
-            k if k == SyntaxKind::NullKeyword as u16 => Some(tsz_solver::types::TypeId::NULL),
-            k if k == syntax_kind_ext::PREFIX_UNARY_EXPRESSION
-                && self.is_negative_literal(node) =>
-            {
-                let prefix = self.arena.get_unary_expr(node)?;
-                let operand = self.arena.get(prefix.operand)?;
-                let literal = self.arena.get_literal(operand)?;
-                if operand.kind == SyntaxKind::BigIntLiteral as u16 {
-                    Some(interner.literal_bigint_with_sign(true, &literal.text.replace('_', "")))
-                } else {
-                    let value = literal
-                        .value
-                        .or_else(|| literal.text.replace('_', "").parse::<f64>().ok())?;
-                    Some(interner.literal_number(-value))
-                }
-            }
-            _ => None,
-        }
-    }
-
-    /// Resolve the element type delegated by a `yield* <iterable>` expression.
-    ///
-    /// The element type is the iterable's iteration (`yield`) type, e.g. `number`
-    /// for `yield* [1, 2]`. Resolution goes through the solver's iterator
-    /// protocol (`get_iterator_info`), which covers array and tuple operands
-    /// structurally; async generators consult the async iterator protocol first
-    /// and fall back to the sync one (mirroring how `yield*` is checked). Returns
-    /// `None` whenever the element type cannot be resolved (or widens to `any`),
-    /// so the caller keeps its conservative `Generator<any, …>` fallback rather
-    /// than emitting a wrong type.
-    fn yield_star_delegated_yield_type_id(
-        &self,
-        expression: NodeIndex,
-        is_async: bool,
-    ) -> Option<tsz_solver::types::TypeId> {
-        let interner = self.type_interner?;
-        let iterable_type = self.get_node_type_or_names(&[expression])?;
-        let element = if is_async {
-            // `get_async_iterable_element_type` tries the async iterator protocol,
-            // then the sync protocol, and returns `any` when neither resolves.
-            tsz_solver::operations::get_async_iterable_element_type(interner, iterable_type)
-        } else {
-            tsz_solver::operations::get_iterator_info(interner, iterable_type, false)
-                .map(|info| info.yield_type)?
-        };
-        let element = self.widen_unique_symbol_value_type_for_dts(element, 0);
-        if element == tsz_solver::types::TypeId::ANY || element == tsz_solver::types::TypeId::ERROR
-        {
-            return None;
-        }
-        Some(element)
     }
 
     pub(in crate::declaration_emitter) fn evaluated_literal_return_type_text_for_returned_identifier(
